requirements.txt
# Minimal requirements for the AI Facial Emotion Analysis System
torch>=1.13.0
torchvision>=0.14.0
opencv-python>=4.6.0
pillow>=9.0.0
onnxruntime>=1.12.0
onnx>=1.12.0
numpy>=1.21.0
matplotlib>=3.5.0
tqdm>=4.64.0
